// Angular-Class11-http-forms-movies-app/movies-app/src/app/feature/movies/models/movie-model.ts
export interface Movie {
  id: string;
  title: string;
  director: string;
  year: number;
  author: string;
  rating: number;
  text: string;
  genres: string;
  likeCount: number;
}

export interface ReviewFormValue {
  title: string;
  director: string;
  year: number;
  genres: string;
  author: string;
  rating: number;
  text: string;
}